include/libproto.h, include/ob_methods.h, include/typesproto.h,   common/ob_methods.c, types/common/describe.c, types/legacy/legacy_describe.c:   Use size_t when working with buffer sizes. include/sproto.h, server/attack.c, server/spell_util.c: Attack type should be   uint32 and not int.
git-svn-id: svn://svn.code.sf.net/p/crossfire/code/server/trunk@11388 282e977c-c81d-0410-88c4-b93c2d0d6712

+++ common/ob_methods.c
@@ -99,7 +99,7 @@ method_ret ob_process(object *op) {
  * @return
  * buf.
  */
-char *ob_describe(const object *op, const object *observer, char *buf, int size) {
+char *ob_describe(const object *op, const object *observer, char *buf, size_t size) {
     ob_methods *methods;
 
     for (methods = &type_methods[op->type]; methods; methods = methods->fallback) {
